feat(pro): lift review derivation and pane out of the Apache packages
review-model + comment-anchors implementations and the React review pane
move to @docx-editor.dev/pro; core keeps the review vocabulary and pure
helpers (layout/review-support) and receives the derivation through the
EditorModule seam. Review tests move to pro; free adapter loses
DocxEditor.Review and gains the rail/Slot/locale integration exports the
pro pane composes with.
