+++ packages/core/src/editor/__tests__/review-test-module.ts
@@ -1,21 +1,19 @@
-// The review module the CORE tests register.
+// A STUB review module for core's seam tests.
 //
-// Review functionality is module-gated: a bare `createDocxEditor` is the free
-// tier — final-state rendering, refused review writes — and the tests in this
-// tree that exercise suggesting, the queue, and the pane are testing the
-// MODULE-REGISTERED behavior. This helper is the in-tree stand-in for the pro
-// package's `reviewModule()`; when the derivation physically moves there, these
-// tests move with it and import the real one.
+// Core tests exercise the seam's mechanics — display-mode gating, command
+// enablement, chrome reasons — which depend on a review contribution being
+// REGISTERED, not on what it derives. The real derivation lives in
+// `@docx-editor.dev/pro` (a package core must not depend on, even in tests);
+// its integration tests live there and use the real `reviewModule()`.
 
 import type { EditorModule } from '../../contracts/modules.ts';
-import { collectReviewItems } from '../../layout/review-model.ts';
 
-export function testReviewModule(): EditorModule {
+export function stubReviewModule(): EditorModule {
   return {
     id: 'review',
     review: {
       displayModes: ['all-markup', 'proposed', 'original'],
-      collectReviewItems,
+      collectReviewItems: () => [],
     },
   };
 }
