+++ packages/pro/package.json
@@ -31,10 +31,14 @@
     "@docx-editor.dev/i18n": "workspace:*"
   },
   "peerDependencies": {
+    "@docx-editor.dev/react": "workspace:*",
     "react": "^18.0.0 || ^19.0.0",
     "react-dom": "^18.0.0 || ^19.0.0"
   },
   "peerDependenciesMeta": {
+    "@docx-editor.dev/react": {
+      "optional": true
+    },
     "react": {
       "optional": true
     },
@@ -43,8 +47,13 @@
     }
   },
   "devDependencies": {
+    "@docx-editor.dev/react": "workspace:*",
+    "@happy-dom/global-registrator": "^20.8.3",
+    "@testing-library/react": "^16.0.0",
     "@types/react": "^19.0.0",
     "fflate": "^0.8.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "typescript": "^5.3.3"
   },
   "keywords": [
