+++ packages/pro/src/__tests__/review-facade.test.ts
@@ -10,9 +10,9 @@ if (!GlobalRegistrator.isRegistered) GlobalRegistrator.register();
 
 import { describe, expect, test } from 'bun:test';
 import { zipSync, strToU8 } from 'fflate';
-import { createDocxEditor, type DocxEditorInstance } from '../docx-editor.ts';
-import { paragraphTextOf } from '../../store/store/tree-ops.ts';
-import { testReviewModule } from './review-test-module.ts';
+import { createDocxEditor, type DocxEditorInstance } from '@docx-editor.dev/core-contract/editor';
+import { paragraphTextOf } from '@docx-editor.dev/core-contract/store';
+import { reviewModule as testReviewModule } from '../review/review-module.ts';
 
 const W = 'http://schemas.openxmlformats.org/wordprocessingml/2006/main';
 const W15 = 'http://schemas.microsoft.com/office/word/2012/wordml';
