+++ packages/react/src/components/DocxEditor.tsx
@@ -15,7 +15,6 @@ import { DocxEditorDocumentOutline } from '../editor/DocxEditorOutline';
 import { Navigation as DocxEditorNavigationCompound } from '../editor/navigation';
 import { DocxEditorPageSetupDialog } from '../editor/DocxEditorPageSetup';
 import { DocxEditorPageNumber, PageNumberTranslationContext } from '../editor/DocxEditorPageNumber';
-import { DocxEditorReview } from '../editor/DocxEditorReview';
 import { DocxEditorHeaderFooterChrome } from '../editor/DocxEditorHeaderFooter';
 import { DocxEditorHyperLink } from '../editor/DocxEditorHyperLink';
 import { DocxEditorNotesChrome } from '../editor/DocxEditorNotes';
@@ -363,11 +362,6 @@ export interface DocxEditorNamespace extends ForwardRefExoticComponent<
    */
   readonly HyperLink: typeof DocxEditorHyperLink;
   /**
-   * The review rail — tracked changes and comments as cards beside the page, with accept,
-   * reject and reply. Place it inside the Viewport, beside `DocxEditor.Content`.
-   */
-  readonly Review: typeof DocxEditorReview;
-  /**
    * The right-click menu over the painted document, with its rows as statics (`.Cut`,
    * `.Copy`, `.Paste`, `.Delete`, `.SelectAll`, `.Item`, `.Slot`, `.Submenu`, …). Mounted
    * by default inside the viewport; `contextMenu={false}` removes it and lets the
@@ -398,7 +392,6 @@ export const DocxEditor: DocxEditorNamespace = Object.assign(DocxEditorImpl, {
   HeaderFooterChrome: DocxEditorHeaderFooterChrome,
   NotesChrome: DocxEditorNotesChrome,
   HyperLink: DocxEditorHyperLink,
-  Review: DocxEditorReview,
   ContextMenu: DocxEditorContextMenuCompound,
   ContentControl: DocxEditorContentControl,
 });
